extras/basic_chat/front/requirements.txt
streamlit==1.35.0
requests==2.32.3
openai>=1.42.0
azure-identity>=1.17.1
python-dotenv>=1.0.1

